Pin puppetlabs-concat to 1.2.1 in fixtures
Puppet Labs is in the process of releasing a 2.0.0 version of
puppetlabs-concat, which will have many breaking changes that affect
many dependent modules. Changes have been merged to master, but not yet
released. While we figure out what changes we need to make in our own
modules and wait for updates in dependent modules, we should just pin
to a stable version.

Change-Id: Ia3c7c528e9ac3c83c38664f2dfab4debd90b5c5d